refactor: extract functions to common
